requirements.txt
PyMySQL==1.0.2
PyJWT==2.6.0
APScheduler==3.9.1
Django==3.2.16
django-simpleui==2022.6.131
django-redis==5.2.0
django-haystack==3.2.1
django-ckeditor~=6.3.2
Whoosh==2.7.4
jieba==0.42.1
requests==2.28.1
Pillow==8.4.0
gevent==22.10.2
qiniu==7.10.0
gunicorn==20.1.0
requests
